+++ CHANGELOG.md
@@ -2286,6 +2286,8 @@ route using fields like `http.path` even for stream routes. This is no longer al
   Previously, the `header_type` was hardcoded to `preserve`, now it can be set to one of the
   following values: `preserve`, `ignore`, `b3`, `b3-single`, `w3c`, `jaeger`, `ot`.
   [#10620](https://github.com/Kong/kong/pull/10620)
+- **StatsD**: allow using credential id as consumer id
+  [#12412](https://github.com/Kong/kong/pull/12412)
 
 #### PDK